SystemConfig: allow app-link to be set to any possible state
Allow app-link entries in /system/etc/sysconfig to set a system
apps app-link state to any of the available states, and not just
"always" as was previously possible.

If the "state" attribute is not present in an app-link
declaration, the state will default to "always" to preserve
existing behavior.

This change is needed to allow certain apps to be set to the
"always-ask" state so they'll be sure to participate in intent
resolution via the disambugation dialog in the default
configuration.

These setting can be changed by the user via the standard
app-link configuration section in Settings -> Apps.

Change-Id: I2f98c69ebbeb40492d7ae8d33327986675b18b65